lapida/requirements.txt
asgiref==3.5.2
Babel==2.9.1
beautifulsoup4==4.9.3
black==22.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.1.0
colorama==0.4.4
cryptography==3.4.6
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-cron==0.6.0
django-import-export==2.5.0
django-phonenumber-field==5.0.0
django-tables2==2.3.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.0.1
flake8==4.0.1
idna==2.10
install==1.3.4
jdcal==1.4.1
MarkupPy==1.14
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
pathspec==0.9.0
phonenumbers==8.12.16
Pillow==9.0.1
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
soupsieve==2.2
sqlparse==0.4.1
sweetify==2.0.0
tablib==3.0.0
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2022.2
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0